# rest/src/main/resources/io/jmix/rest/rest-openapi.yaml
openapi: 3.0.1
info:
  title: Jmix Generic REST API
  description: |2

    ## Overview

    Jmix Generic REST API provides the following functionality:

    * Entities CRUD operations
    * Execution of predefined JPQL queries
    * Execution of exposed service methods
    * Getting metadata (entities, fetch plans, enumerations, datatypes)
    * Getting localized messages
    * Getting current user permissions (roles, access to entities and attributes, specific policies)
    * Getting attributes of the current user
    * Working with files

    The REST API is available at the `/rest/` URL, e.g. `http://localhost:8080/rest/entities/Order`

    ## Authentication

    ### Authentication using the OAuth2 protocol

    Jmix REST API uses an OAuth2 resource owner password credentials grant to protect the endpoints (https://datatracker.ietf.org/doc/html/rfc6749#section-4.3).

    Before accessing a protected resource, a user must get an authentication token. The token can be obtained at the `/oauth/token` URL.

    The `/oauth/token` endpoint is protected with the basic authentication. The request must contain the `Authorization` header for basic
    authentication with REST API client credentials. Please note that client credentials are not the same as the user credentials.
    Client means an application that uses the REST API. The default client credentials are defined by the following
...
    The request format must be `application/x-www-form-urlencoded`.

...

...

    ```
...
    ```

...

...

...
servers:
...
security:
  - oauth2:
      - rest-api
...
paths:
...
  /queries/{entityName}:
    get:
      tags:
        - Queries
      summary: Get a list of queries
      description: |
        Gets a list of predefined queries for the entity
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/predefinedQuery'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found.
          content: {}
  /queries/{entityName}/{queryName}:
    get:
      tags:
        - Queries
      summary: Execute a query
      description: |
        Executes a predefined query. Query parameters values must be defined in the request parameters with the same name.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: queryName
          in: path
          description: Query name
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Number of extracted entities
          schema:
            type: string
        - name: offset
          in: query
          description: Position of the first result to retrieve
          schema:
            type: string
        - name: fetchPlan
          in: query
          description: |
            Name of the fetch plan which is used for loading the entity. Specify this parameter if you want to extract entities with the fetch plan
            other than it is defined in the REST queries configuration file.
          schema:
            type: string
        - name: returnNulls
          in: query
          description: Specifies whether null fields will be written to the result JSON
          schema:
            type: boolean
        - name: returnCount
          in: query
          description: Specifies whether the total count of entities should be returned
            in the 'X-Total-Count' header
          schema:
            type: boolean
        - name: dynamicAttributes
          in: query
          description: Specifies whether entity dynamic attributes should be returned
          schema:
            type: boolean
      responses:
        200:
          description: Success. A list of entities is returned in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or query with the given name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
        - Queries
      summary: Execute a query
      description: |
        Executes a predefined query. Query parameters must be passed in the request body as JSON map.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: queryName
          in: path
          description: Query name
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Number of extracted entities
          schema:
            type: string
        - name: offset
          in: query
          description: Position of the first result to retrieve
          schema:
            type: string
        - name: fetchPlan
          in: query
          description: |
            Name of the fetch plan which is used for loading the entity. Specify this parameter if you want to extract entities with the fetch plan
            other than it is defined in the REST queries configuration file.
          schema:
            type: string
        - name: returnNulls
          in: query
          description: Specifies whether null fields will be written to the result JSON
          schema:
            type: boolean
        - name: returnCount
          in: query
          description: Specifies whether the total count of entities should be returned
            in the 'X-Total-Count' header
          schema:
            type: boolean
        - name: dynamicAttributes
          in: query
          description: Specifies whether entity dynamic attributes should be returned
          schema:
            type: boolean
      responses:
        200:
          description: Success. A list of entitie is returned in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or query with the given name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /queries/{entityName}/{queryName}/count:
    get:
      tags:
        - Queries
      summary: Return a number of entities in query result
      description: |
        Returns a number of entities that matches the query. You can use the `all` keyword for the `queryNameParam`
        to get the number of all available entities.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: queryName
          in: path
          description: Query name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                title: Count
                type: integer
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or query with the given name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
        - Queries
      summary: Return a number of entities in query result
      description: |
        Returns a number of entities that matches the query. You can use the `all` keyword for the `queryNameParam`
        to get the number of all available entities.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: queryName
          in: path
          description: Query name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                title: Count
                type: integer
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or query with the given name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
...
    error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error description
...
    entityFull:
      required:
        - _entityName
        - id
      type: object
      properties:
        id:
          type: string
          example: 4d8ef1d4-90b6-c075-8bfd-79181ff889ef
        _entityName:
          type: string
          description: Entity name
          example: sec$User
        _instanceName:
          type: string
          description: Entity instance name. This value must be used for displaying
            the entity in list.
          example: Smith [smith]
        otherEntityFields:
          type: string
          description: There must be a set of entity fields instead of this one here.
          example: There must be a set of entity fields instead of this one here
    predefinedQuery:
      type: object
      properties:
        name:
          type: string
          description: Query name
          example: usersByLastName
        jpql:
          type: string
          description: JPQL query
          example: select u from sec$User u where u.lastName = :lastName
        entityName:
          type: string
          description: Entity name
          example: sec$User
        fetchPlanName:
          type: string
          description: Fetch plan name
          example: user.edit
        params:
          type: object
          properties:
            name:
              type: string
              description: Parameter name
              example: lastName
            type:
              type: string
              description: Parameter type. The fully qualified java class name or
                primitive type name is returned here.
              example: java.lang.String
          description: Collection of query parameters
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: http://localhost:8080/rest/oauth/token
          scopes:
            rest-api: use REST API
    tokenEndpointBasic:
      type: http
      scheme: basic